+++ internal/handlers/json_errors_test.go
