+++ internal/storage/entry.go
