+++ internal/storage/entry_test.go
