+++ internal/zipper/writer.go
