+++ internal/zipper/writer_test.go
